+++ gdsfactory/functions.py
@@ -4,6 +4,7 @@ from functools import lru_cache
 from gdsfactory.cell import cell
 from gdsfactory.component import Component
 from gdsfactory.port import auto_rename_ports
+from gdsfactory.types import Optional
 
 cache = lru_cache
 
@@ -37,6 +38,22 @@ def rotate(
     return component_new
 
 
+@cell
+def move(
+    component: Component,
+    origin=(0, 0),
+    destination=None,
+    axis: Optional[str] = None,
+) -> Component:
+    component_new = Component()
+    component_new.component = component
+    ref = component_new.add_ref(component)
+    ref.move(origin=origin, destination=destination, axis=axis)
+    component_new.add_ports(ref.ports)
+    component_new.copy_child_info(component)
+    return component_new
+
+
 def update_info(component: Component, **kwargs) -> Component:
     """Returns Component with updated info."""
     component.info.update(**kwargs)
@@ -52,6 +69,9 @@ if __name__ == "__main__":
     c = gf.components.mmi1x2()
     cr = c.rotate()
     cr.show()
+
+    cm = move(c, destination=(20, 20))
+    cm.show()
 
     # cr = rotate(component=c)
     # cr.show()