+++ gdsfactory/geometry/boolean.py
@@ -23,6 +23,10 @@ def boolean(
     Note that 'A+B' is equivalent to 'or', 'A-B' is equivalent to 'not', and
     'B-A' is equivalent to 'not' with the operands switched
 
+    gdsfactory wrapper for phidl.geometry.boolean
+
+    You can also use gdsfactory.drc.boolean that uses Klayout backend
+
     Args:
         A: Component(/Reference) or list of Component(/References)
         B: Component(/Reference) or list of Component(/References)